feat: migrate h&b to use tree structure
